Newton's Quest: Unlock the Alien Ship
Newton, the alien hunter, is on a mission in a new movie released on Newflix called Newton's Quest. In one scene, Newton finds an alien ship built with futuristic technology. To take control of the ship, he needs a digital key that consists of a pattern of three bits: a, b, and c.

The ship can only be unlocked if these three bits, when evaluated as a decimal number, result in a power of 2.

Your task: You are given three space-separated bits a, b, and c. These bits represent a binary number. Your goal is to determine if this binary number, when converted to decimal, is a power of 2 or not. If it is then you need to print "Yes! Ship Unlocked." otherwise print "No! Better luck next time."

Example:
If the input is 0 1 1, the decimal equivalent is calculated as:
0 * 22 + 1 * 21 + 1 * 20 = 3, which is not a power of 2.

Note: Refer Input section, output section and example test cases.

 
Input
A single line containing three space-separated integers a, b, and c (each either 0 or 1).
Output
Print "Yes! Ship Unlocked." if the decimal number is a power of 2. Otherwise, Print "No! Better luck next time."